honor spring.dashboard.openUrl on opening endpoints
